Fixed a bug in the UI item parser, were some values should have been read as floats rather than ints.
Put in 640/480 positioning on text items.

+++ idLib/Engine/Content/ui/ui_item.cs
@@ -107,9 +107,9 @@ namespace idLib.Engine.Content.ui
             font = reader.ReadString();
             fontSize = reader.ReadInt32();
             textalignment = reader.ReadInt32();
-            textalignx = reader.ReadInt32();
-            textaligny = reader.ReadInt32();
-            textscale = reader.ReadInt32();
+            textalignx = reader.ReadSingle();
+            textaligny = reader.ReadSingle();
+            textscale = reader.ReadSingle();
             textStyle = reader.ReadInt32();
             text = reader.ReadString();
             textSavegameInfo = reader.ReadBoolean();
